ALSA: hda - hdmi audio add support of switch
This patch adds the support of the switch class for android.

This is a simple implementation for the hdmi monitor switch.
In this patch, the hdmi_switch status represents all the pins
monitor connection status. This means if there is one monitor
connected to any pin, the hdmi_switch will be 1.

Signed-off-by: Libin Yang <libin.yang@intel.com>